Add a basic turn change logic.
This is implemented by setting a message when end turn is clicked. It will set a signal to inform that the turn has been changed.
The heartbeat will be updated each second, so when it is informed that the turn has changed, it will call the private method to process it.

Currently no actions after turn changes has been implemented, only a notification showing on Player 1's side([^1]).

[^1]: The notification API is only available for Player 1 in basic commands, so all the notifications show on Player 1's side for now.