Add two deck classes and draw card logic.
This is probably not a Card drawing logic because it just loads the card images to an ArrayList, without considering any of the card config stuff.

The deck classes will initialise all the cards and get them shuffled. This class provides a drawCard function. When we need to draw card (like in GameState and Initalize), we will call this function and it will serve us the next card in the deck and remove it from the deck.

In the EndTurnClicked (and Initalize), I have provided a way to draw card with DrawCardCommandBuilder.