Create UnitFactory. Please use it to create Units from this on.
From this commit on, please always create Units with UnitFactory.

This UnitFactory includes a counter to assign ID for every generated Units. This will prevent it from conflicting with each other. When the ID conflicts, it may influence the moving of elements.

This commit only includes Avatars. Other configs coming.