Add logic to prevent units from moving if a unit is moving.
If you move another unit (B) when a unit (A) is moving, A will be stuck on the track as B moves.
The added logic supports locking the movement highlights when we start the moving and set it to unlocked when the movement is done.

Notice this commit only locks the movement highlights. It depends on whether this frontend issue appears when dealing a card. If yes, we will need to lock that too.